move transaction apply methods into their own package
