+++ data/transactions/keyreg.go
@@ -17,8 +17,6 @@
 package transactions
 
 import (
-	"fmt"
-
 	"github.com/algorand/go-algorand/crypto"
 	"github.com/algorand/go-algorand/data/basics"
 )
@@ -33,54 +31,4 @@ type KeyregTxnFields struct {
 	VoteLast         basics.Round                    `codec:"votelst"`
 	VoteKeyDilution  uint64                          `codec:"votekd"`
 	Nonparticipation bool                            `codec:"nonpart"`
-}
-
-// Apply changes the balances according to this transaction.
-func (keyreg KeyregTxnFields) apply(header Header, balances Balances, spec SpecialAddresses, ad *ApplyData) error {
-	if header.Sender == spec.FeeSink {
-		return fmt.Errorf("cannot register participation key for fee sink's address %v ", header.Sender)
-	}
-
-	// Get the user's balance entry
-	record, err := balances.Get(header.Sender, false)
-	if err != nil {
-		return err
-	}
-
-	// non-participatory accounts cannot be brought online (or offline)
-	if record.Status == basics.NotParticipating {
-		return fmt.Errorf("cannot change online/offline status of non-participating account %v", header.Sender)
-	}
-
-	// Update the registered keys and mark account as online
-	// (or, if the voting or selection keys are zero, offline/not-participating)
-	record.VoteID = keyreg.VotePK
-	record.SelectionID = keyreg.SelectionPK
-	if (keyreg.VotePK == crypto.OneTimeSignatureVerifier{} || keyreg.SelectionPK == crypto.VRFVerifier{}) {
-		if keyreg.Nonparticipation {
-			if balances.ConsensusParams().SupportBecomeNonParticipatingTransactions {
-				record.Status = basics.NotParticipating
-			} else {
-				return fmt.Errorf("transaction tries to mark an account as nonparticipating, but that transaction is not supported")
-			}
-		} else {
-			record.Status = basics.Offline
-		}
-		record.VoteFirstValid = 0
-		record.VoteLastValid = 0
-		record.VoteKeyDilution = 0
-	} else {
-		record.Status = basics.Online
-		record.VoteFirstValid = keyreg.VoteFirst
-		record.VoteLastValid = keyreg.VoteLast
-		record.VoteKeyDilution = keyreg.VoteKeyDilution
-	}
-
-	// Write the updated entry
-	err = balances.Put(record)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
