+++ data/transactions/payment.go
@@ -53,71 +53,3 @@ func (payment PaymentTxnFields) checkSpender(header Header, spec SpecialAddresse
 	}
 	return nil
 }
-
-// Apply changes the balances according to this transaction.
-// The ApplyData argument should reflect the changes made by
-// apply().  It may already include changes made by the caller
-// (i.e., Transaction.Apply), so apply() must update it rather
-// than overwriting it.  For example, Transaction.Apply() may
-// have updated ad.SenderRewards, and this function should only
-// add to ad.SenderRewards (if needed), but not overwrite it.
-func (payment PaymentTxnFields) apply(header Header, balances Balances, spec SpecialAddresses, ad *ApplyData) error {
-	// move tx money
-	if !payment.Amount.IsZero() || payment.Receiver != (basics.Address{}) {
-		err := balances.Move(header.Sender, payment.Receiver, payment.Amount, &ad.SenderRewards, &ad.ReceiverRewards)
-		if err != nil {
-			return err
-		}
-	}
-
-	if payment.CloseRemainderTo != (basics.Address{}) {
-		rec, err := balances.Get(header.Sender, true)
-		if err != nil {
-			return err
-		}
-
-		closeAmount := rec.AccountData.MicroAlgos
-		ad.ClosingAmount = closeAmount
-		err = balances.Move(header.Sender, payment.CloseRemainderTo, closeAmount, &ad.SenderRewards, &ad.CloseRewards)
-		if err != nil {
-			return err
-		}
-
-		// Confirm that we have no balance left
-		rec, err = balances.Get(header.Sender, true)
-		if !rec.AccountData.MicroAlgos.IsZero() {
-			return fmt.Errorf("balance %d still not zero after CloseRemainderTo", rec.AccountData.MicroAlgos.Raw)
-		}
-
-		// Confirm that there is no asset-related state in the account
-		if len(rec.Assets) > 0 {
-			return fmt.Errorf("cannot close: %d outstanding assets", len(rec.Assets))
-		}
-
-		if len(rec.AssetParams) > 0 {
-			// This should be impossible because every asset created
-			// by an account (in AssetParams) must also appear in Assets,
-			// which we checked above.
-			return fmt.Errorf("cannot close: %d outstanding created assets", len(rec.AssetParams))
-		}
-
-		// Confirm that there is no application-related state remaining
-		if len(rec.AppLocalStates) > 0 {
-			return fmt.Errorf("cannot close: %d outstanding applications opted in. Please opt out or clear them", len(rec.AppLocalStates))
-		}
-
-		// Can't have created apps remaining either
-		if len(rec.AppParams) > 0 {
-			return fmt.Errorf("cannot close: %d outstanding created applications", len(rec.AppParams))
-		}
-
-		// Clear out entire account record, to allow the DB to GC it
-		rec.AccountData = basics.AccountData{}
-		err = balances.Put(rec)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
