+++ data/transactions/transaction.go
@@ -46,41 +46,6 @@ type SpecialAddresses struct {
 	RewardsPool basics.Address
 }
 
-// Balances allow to move MicroAlgos from one address to another and to update balance records, or to access and modify individual balance records
-// After a call to Put (or Move), future calls to Get or Move will reflect the updated balance record(s)
-type Balances interface {
-	// Get looks up the balance record for an address
-	// If the account is known to be empty, then err should be nil and the returned balance record should have the given address and empty AccountData
-	// withPendingRewards specifies whether pending rewards should be applied.
-	// A non-nil error means the lookup is impossible (e.g., if the database doesn't have necessary state anymore)
-	Get(addr basics.Address, withPendingRewards bool) (basics.BalanceRecord, error)
-
-	Put(basics.BalanceRecord) error
-
-	// PutWithCreatable is like Put, but should be used when creating or deleting an asset or application.
-	PutWithCreatable(record basics.BalanceRecord, newCreatable *basics.CreatableLocator, deletedCreatable *basics.CreatableLocator) error
-
-	// GetCreator gets the address of the account that created a given creatable
-	GetCreator(cidx basics.CreatableIndex, ctype basics.CreatableType) (basics.Address, bool, error)
-
-	// Move MicroAlgos from one account to another, doing all necessary overflow checking (convenience method)
-	// TODO: Does this need to be part of the balances interface, or can it just be implemented here as a function that calls Put and Get?
-	Move(src, dst basics.Address, amount basics.MicroAlgos, srcRewards *basics.MicroAlgos, dstRewards *basics.MicroAlgos) error
-
-	// Balances correspond to a Round, which mean that they also correspond
-	// to a ConsensusParams.  This returns those parameters.
-	ConsensusParams() config.ConsensusParams
-}
-
-// StateEvaluator is an interface that provides some Stateful TEAL
-// functionality that may be passed through to Apply from ledger, avoiding a
-// circular dependency between the logic and transactions packages
-type StateEvaluator interface {
-	Eval(program []byte) (pass bool, stateDelta basics.EvalDelta, err error)
-	Check(program []byte) (cost int, err error)
-	InitLedger(balances Balances, appIdx basics.AppIndex, schemas basics.StateSchemas) error
-}
-
 // Header captures the fields common to every transaction type.
 type Header struct {
 	_struct struct{} `codec:",omitempty,omitemptyarray"`
@@ -531,71 +496,6 @@ func (tx Transaction) EstimateEncodedSize() int {
 		Sig: crypto.Signature{1},
 	}
 	return stx.GetEncodedLength()
-}
-
-// Apply changes the balances according to this transaction.
-func (tx Transaction) Apply(balances Balances, steva StateEvaluator, spec SpecialAddresses, ctr uint64) (ad ApplyData, err error) {
-	params := balances.ConsensusParams()
-
-	// move fee to pool
-	err = balances.Move(tx.Sender, spec.FeeSink, tx.Fee, &ad.SenderRewards, nil)
-	if err != nil {
-		return
-	}
-
-	// rekeying: update balrecord.AuthAddr to tx.RekeyTo if provided
-	if (tx.RekeyTo != basics.Address{}) {
-		var record basics.BalanceRecord
-		record, err = balances.Get(tx.Sender, false)
-		if err != nil {
-			return
-		}
-		// Special case: rekeying to the account's actual address just sets balrecord.AuthAddr to 0
-		// This saves 32 bytes in your balance record if you want to go back to using your original key
-		if tx.RekeyTo == tx.Sender {
-			record.AuthAddr = basics.Address{}
-		} else {
-			record.AuthAddr = tx.RekeyTo
-		}
-
-		err = balances.Put(record)
-		if err != nil {
-			return
-		}
-	}
-
-	switch tx.Type {
-	case protocol.PaymentTx:
-		err = tx.PaymentTxnFields.apply(tx.Header, balances, spec, &ad)
-
-	case protocol.KeyRegistrationTx:
-		err = tx.KeyregTxnFields.apply(tx.Header, balances, spec, &ad)
-
-	case protocol.AssetConfigTx:
-		err = tx.AssetConfigTxnFields.apply(tx.Header, balances, spec, &ad, ctr)
-
-	case protocol.AssetTransferTx:
-		err = tx.AssetTransferTxnFields.apply(tx.Header, balances, spec, &ad)
-
-	case protocol.AssetFreezeTx:
-		err = tx.AssetFreezeTxnFields.apply(tx.Header, balances, spec, &ad)
-
-	case protocol.ApplicationCallTx:
-		err = tx.ApplicationCallTxnFields.apply(tx.Header, balances, spec, &ad, ctr, steva)
-
-	default:
-		err = fmt.Errorf("Unknown transaction type %v", tx.Type)
-	}
-
-	// If the protocol does not support rewards in ApplyData,
-	// clear them out.
-	if !params.RewardsInApplyData {
-		ad.SenderRewards = basics.MicroAlgos{}
-		ad.ReceiverRewards = basics.MicroAlgos{}
-		ad.CloseRewards = basics.MicroAlgos{}
-	}
-
-	return
 }
 
 // TxnContext describes the context in which a transaction can appear
