+++ ledger/eval.go
@@ -29,6 +29,7 @@ import (
 	"github.com/algorand/go-algorand/data/transactions"
 	"github.com/algorand/go-algorand/data/transactions/logic"
 	"github.com/algorand/go-algorand/data/transactions/verify"
+	"github.com/algorand/go-algorand/ledger/apply"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/protocol"
 	"github.com/algorand/go-algorand/util/execpool"
@@ -78,7 +79,7 @@ func (x *roundCowBase) txnCounter() uint64 {
 	return x.txnCount
 }
 
-// wrappers for roundCowState to satisfy the (current) transactions.Balances interface
+// wrappers for roundCowState to satisfy the (current) apply.Balances interface
 func (cs *roundCowState) Get(addr basics.Address, withPendingRewards bool) (basics.BalanceRecord, error) {
 	acctdata, err := cs.lookup(addr)
 	if err != nil {
@@ -478,7 +479,7 @@ func (eval *BlockEvaluator) prepareAppEvaluators(txgroup []transactions.SignedTx
 		}
 
 		// Construct an appTealEvaluator (implements
-		// transactions.StateEvaluator) for use in ApplicationCall transactions.
+		// apply.StateEvaluator) for use in ApplicationCall transactions.
 		steva := appTealEvaluator{
 			evalParams: logic.EvalParams{
 				Txn:        &groupNoAD[i],
@@ -614,7 +615,7 @@ func (eval *BlockEvaluator) transaction(txn transactions.SignedTxn, appEval *app
 	}
 
 	// Apply the transaction, updating the cow balances
-	applyData, err := txn.Txn.Apply(cow, appEval, spec, cow.txnCounter())
+	applyData, err := applyTransaction(txn.Txn, cow, appEval, spec, cow.txnCounter())
 	if err != nil {
 		return fmt.Errorf("transaction %v: %v", txid, err)
 	}
@@ -682,6 +683,71 @@ func (eval *BlockEvaluator) transaction(txn transactions.SignedTxn, appEval *app
 	cow.addTx(txn.Txn, txid)
 
 	return nil
+}
+
+// applyTransaction changes the balances according to this transaction.
+func applyTransaction(tx transactions.Transaction, balances apply.Balances, steva apply.StateEvaluator, spec transactions.SpecialAddresses, ctr uint64) (ad transactions.ApplyData, err error) {
+	params := balances.ConsensusParams()
+
+	// move fee to pool
+	err = balances.Move(tx.Sender, spec.FeeSink, tx.Fee, &ad.SenderRewards, nil)
+	if err != nil {
+		return
+	}
+
+	// rekeying: update balrecord.AuthAddr to tx.RekeyTo if provided
+	if (tx.RekeyTo != basics.Address{}) {
+		var record basics.BalanceRecord
+		record, err = balances.Get(tx.Sender, false)
+		if err != nil {
+			return
+		}
+		// Special case: rekeying to the account's actual address just sets balrecord.AuthAddr to 0
+		// This saves 32 bytes in your balance record if you want to go back to using your original key
+		if tx.RekeyTo == tx.Sender {
+			record.AuthAddr = basics.Address{}
+		} else {
+			record.AuthAddr = tx.RekeyTo
+		}
+
+		err = balances.Put(record)
+		if err != nil {
+			return
+		}
+	}
+
+	switch tx.Type {
+	case protocol.PaymentTx:
+		err = apply.Payment(tx.PaymentTxnFields, tx.Header, balances, spec, &ad)
+
+	case protocol.KeyRegistrationTx:
+		err = apply.Keyreg(tx.KeyregTxnFields, tx.Header, balances, spec, &ad)
+
+	case protocol.AssetConfigTx:
+		err = apply.AssetConfig(tx.AssetConfigTxnFields, tx.Header, balances, spec, &ad, ctr)
+
+	case protocol.AssetTransferTx:
+		err = apply.AssetTransfer(tx.AssetTransferTxnFields, tx.Header, balances, spec, &ad)
+
+	case protocol.AssetFreezeTx:
+		err = apply.AssetFreeze(tx.AssetFreezeTxnFields, tx.Header, balances, spec, &ad)
+
+	case protocol.ApplicationCallTx:
+		err = apply.ApplicationCall(tx.ApplicationCallTxnFields, tx.Header, balances, &ad, ctr, steva)
+
+	default:
+		err = fmt.Errorf("Unknown transaction type %v", tx.Type)
+	}
+
+	// If the protocol does not support rewards in ApplyData,
+	// clear them out.
+	if !params.RewardsInApplyData {
+		ad.SenderRewards = basics.MicroAlgos{}
+		ad.ReceiverRewards = basics.MicroAlgos{}
+		ad.CloseRewards = basics.MicroAlgos{}
+	}
+
+	return
 }
 
 // Call "endOfBlock" after all the block's rewards and transactions are processed.
